# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Демо-проект на Go с поддержкой PostgreSQL и MongoDB",
        "title": "Go Currency API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/": {
            "get": {
                "description": "Проверяет работоспособность сервиса",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Проверка работы API",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.HealthcheckResponse"
                        }
                    }
                }
            }
        },
        "/currencies/latest/": {
            "get": {
                "description": "Получение списка актуальных валют (курса к доллару)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "currencies"
                ],
                "summary": "Получение текущего курса валют",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dtos.Currency"
                            }
                        }
                    }
                }
            }
        },
        "/currencies/pair/": {
            "get": {
                "description": "Получение отношения валюты left к валюте right",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "currencies"
                ],
                "summary": "Получение курса валютной пары",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Left currency type",
                        "name": "left",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Right currency type",
                        "name": "right",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.CurrencyPair"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.HealthcheckResponse": {
            "type": "object",
            "properties": {
                "api_name": {
                    "type": "string"
                },
                "api_version": {
                    "type": "string"
                },
                "app_mode": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "dtos.Currency": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "rate": {
                    "type": "number"
                },
                "type": {
                    "$ref": "#/definitions/dtos.CurrencyType"
                }
            }
        },
        "dtos.CurrencyPair": {
            "type": "object",
            "properties": {
                "left": {
                    "$ref": "#/definitions/dtos.CurrencyType"
                },
                "rate": {
                    "type": "number"
                },
                "right": {
                    "$ref": "#/definitions/dtos.CurrencyType"
                }
            }
        },
        "dtos.CurrencyType": {
            "type": "string",
            "enum": [
                "USD",
                "RUB",
                "EUR",
                "JPY"
            ],
            "x-enum-varnames": [
                "USD",
                "RUB",
                "EUR",
                "JPY"
            ]
        }
    }
}